Generated by Spark: scan around the project and ensure anything that is pointing to anything labeled test has good actual logic behind it then remove test by relocating the file content and zeroing out the test named file or ensuring test is not in the function call name etc - basic pretend population data is the exception here but even that should probably not really exist .
you need to be extremely sure before just deciding to zero out files with test in them every time you have done this - the application breaks with no options but to roll back. maybe do one or two at a time so i can find the problem when you get to this step?